feat: Implement Replay Mode for SlotcraftClient
This commit introduces a new 'Replay Mode' to the SlotcraftClient.

This mode allows the client to be initialized with an HTTP/HTTPS URL pointing to a JSON file. Instead of connecting to a WebSocket server, the client fetches this file and uses its content to simulate a game session. This is highly useful for debugging, integration testing, and UI development without a live backend.

Key changes:
- Refactored the `SlotcraftClient` into a factory pattern.
- Created an `ISlotcraftClientImpl` interface to define the client contract.
- Moved the original WebSocket logic into a `SlotcraftClientLive` class.
- Implemented the new `SlotcraftClientReplay` class to handle fetching and simulating from a JSON file.
- Added the `node-fetch` dependency to support fetching in a Node.js environment.
- Added new tests for the replay mode and updated existing integration tests to accommodate the refactoring.
- Updated project documentation (`jules.md`) to include details about the new feature.
