+++ jules/plan028.md
@@ -1,52 +1,55 @@
-# Plan 028: Cache defaultScene from gamemoduleinfo
+# Plan 028: Implement Replay Mode
 
-## 1. Understanding the Goal
+## 1. Objective
 
-The user wants to cache the `defaultScene` data received in the `gamemoduleinfo` message. This data needs to be transformed from a complex object structure into a simple 2D array and stored in the `UserInfo` object. The transformation logic must be an exported utility function with its own tests and documentation.
+To add a "replay mode" to `SlotcraftClient`. This mode will allow the client to run against a static JSON file (served via a URL) instead of a live WebSocket server. This is useful for debugging, testing, and demonstrating specific game scenarios without requiring a live server connection.
 
-## 2. Task Decomposition
+## 2. Requirements
 
-### Step 1: Create a Utility Function for Data Transformation
-- Create a new file: `src/utils.ts`.
-- Inside this file, implement an exported function named `transformSceneData`.
-- This function will accept an object argument matching the structure of `defaultScene` (i.e., `{ values: [{ values: [...] }] }`).
-- It will return a 2D array of numbers (`number[][]`).
-- Add comprehensive JSDoc comments explaining what the function does, its parameters, and what it returns.
+- The client should be initialized into replay mode if the provided URL is for a JSON file (HTTP/HTTPS).
+- The replay mode should be mutually exclusive with the standard WebSocket mode.
+- The implementation should be clean, separating the logic for the two modes, likely using an interface and two concrete classes.
+- The public API of `SlotcraftClient` should remain unchanged, allowing `examples/example001.ts` to work with either mode by simply changing the URL.
+- User actions in replay mode (like `spin` or `collect`) should resolve immediately with the data from the JSON file.
+- The feature needs to be documented in `jules.md`.
 
-### Step 2: Write Unit Tests for the Utility Function
-- Create a new test file: `tests/utils.test.ts`.
-- Write test cases for `transformSceneData`.
-- Test the happy path with the example data provided in the task description.
-- Test edge cases, such as:
-  - An empty outer `values` array.
-  - An inner `values` array being empty.
-  - Malformed input (e.g., missing `values` properties).
-  - Input being `null` or `undefined`.
+## 3. Implementation Strategy
 
-### Step 3: Update Core Types
-- Open `src/types.ts`.
-- Add a new optional property `defaultScene?: number[][]` to the `UserInfo` interface.
+### Step 1: Create a Client Interface
 
-### Step 4: Integrate the Transformation Logic into the Message Handler
-- Open `src/main.ts`.
-- Import the `transformSceneData` function from `./utils`.
-- Locate the `updateCaches` method and the `case 'gamemoduleinfo':` block.
-- Inside this block, check if `g.defaultScene` exists.
-- If it does, call `transformSceneData(g.defaultScene)` and assign the result to `this.userInfo.defaultScene`.
+- Define a new interface, `ISlotcraftClientImpl`, in `src/types.ts`.
+- This interface will declare all the public methods and properties of the client (`connect`, `enterGame`, `spin`, `collect`, `selectOptional`, `disconnect`, `getState`, `getUserInfo`, `on`, `off`, `once`). This ensures that both the live and replay implementations adhere to the same contract.
 
-### Step 5: Export the New Utility
-- Open `src/index.ts`.
-- Export the `transformSceneData` function from `./utils` to make it accessible to external consumers of the library.
+### Step 2: Refactor Live Client
 
-### Step 6: Verify the Changes
-- Run the full check script: `npm run check`.
-- This will execute linting, all tests (including the new ones), and the build process.
-- Ensure all checks pass without errors.
+- Create a new file `src/live-client.ts`.
+- Move the entire implementation of the current `SlotcraftClient` class from `src/main.ts` into this new file.
+- Rename the class to `SlotcraftClientLive` and make it implement the `ISlotcraftClientImpl` interface.
+- Export the new class.
 
-### Step 7: Update Documentation
-- Update the main project documentation file, `jules.md`.
-- Add a new section, e.g., "## Utilities", to document the `transformSceneData` function. Describe its purpose and usage.
+### Step 3: Implement Replay Client
 
-## 3. Post-Execution
-- After completing the task, create a report file `jules/plan028-report.md` detailing the execution process, challenges, and solutions.
-- Review and update `agents.md` if the changes are relevant for future AI agent work (in this case, it's unlikely to be necessary, but should be checked).
+- Create a new file `src/replay-client.ts`.
+- Create a new class `SlotcraftClientReplay` that implements `ISlotcraftClientImpl`.
+- The constructor will accept the client options, including the JSON URL.
+- The `connect` method will use `fetch` to get the JSON data from the URL and store it. It will simulate the connection and login process, setting the state to `LOGGED_IN`.
+- The `enterGame` method will process the `gmi` part of the JSON and set the state to `IN_GAME` or `SPINEND` as appropriate, mimicking the resume logic.
+- The `spin`, `collect`, and `selectOptional` methods will be simple stubs that immediately return a resolved promise with the relevant data from the loaded JSON, as the entire spin result is already known.
+- All other methods (`getState`, `getUserInfo`, event handlers) will be implemented to manage the state of the replay session.
+
+### Step 4: Update the Main `SlotcraftClient`
+
+- Modify `src/main.ts` to turn `SlotcraftClient` into a factory/wrapper class.
+- The constructor will inspect the `url` option. If it matches a websocket protocol (`ws://` or `wss://`), it will instantiate `SlotcraftClientLive`. If it matches an HTTP protocol (`http://` or `https://`), it will instantiate `SlotcraftClientReplay`.
+- The `SlotcraftClient` will hold a private instance of `ISlotcraftClientImpl` and delegate all public method calls to it.
+
+### Step 5: Add Tests for Replay Mode
+
+- Create a new test file `tests/replay.test.ts`.
+- Add tests to verify that `SlotcraftClientReplay` correctly loads data from a mock JSON URL, initializes its state, and that its methods (`enterGame`, `spin`, etc.) return the expected data.
+
+### Step 6: Finalize and Document
+
+- Run `npm run check` to ensure all tests pass and there are no linting errors.
+- Update `jules.md` with a detailed section about the new Replay Mode.
+- Create the final report `jules/plan028-report.md`.
