+++ package.json
@@ -35,6 +35,7 @@
   },
   "homepage": "https://github.com/zhs007/slotcraft-client-net#readme",
   "devDependencies": {
+    "@eslint/js": "^9.34.0",
     "@types/node": "^20.14.10",
     "@types/ws": "^8.18.1",
     "@typescript-eslint/eslint-plugin": "^8.0.0",
@@ -48,7 +49,10 @@
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
     "vitest": "^3.2.4",
-    "ws": "^8.18.0",
-    "@eslint/js": "^9.34.0"
+    "ws": "^8.18.0"
+  },
+  "dependencies": {
+    "@types/node-fetch": "^2.6.13",
+    "node-fetch": "^2.7.0"
   }
 }
