+++ src/types.ts
@@ -192,3 +192,23 @@ export interface SpinParams {
   // Allow extra ctrlparam fields if needed
   [key: string]: any;
 }
+
+/**
+ * Defines the public interface for a Slotcraft client implementation.
+ * This allows for different implementations (e.g., live WebSocket vs. replay from JSON)
+ * to be used interchangeably.
+ */
+export interface ISlotcraftClientImpl {
+  getState(): ConnectionState;
+  getUserInfo(): Readonly<UserInfo>;
+  connect(token?: string): Promise<void>;
+  enterGame(gamecode?: string): Promise<any>;
+  spin(params: SpinParams): Promise<any>;
+  collect(playIndex?: number): Promise<any>;
+  selectOptional(index: number): Promise<any>;
+  disconnect(): void;
+  send(cmdid: string, params: any): Promise<any>;
+  on(event: string, callback: (...args: any[]) => void): void;
+  off(event: string, callback: (...args: any[]) => void): void;
+  once(event: string, callback: (...args: any[]) => void): void;
+}
